default_requirements.txt
# in image: Django==1.9.1
djangorestframework==3.3.2
gunicorn==19.4.5
# in image: mysqlclient==1.3.7
# in image: psycopg2==2.6.1
